requirements.txt
amqp==1.4.6
anyjson==0.3.3
billiard==3.3.0.20
celery==3.1.18
certifi==2015.4.28
cookies==2.2.1
Django==1.8.3
django-bower==5.0.4
django-guardian==1.3
django-nvd3==0.8.2
funcsigs==0.4
Jinja2==2.8
kombu==3.0.26
lxml==3.4.4
MarkupSafe==0.23
mock==1.2.0
pbr==1.3.0
psycopg2==2.6.1
python-forecastio==1.3.4
python-nvd3==0.13.7
python-slugify==0.0.9
pytz==2015.4
redis==2.10.3
requests==2.7.0
responses==0.4.0
six==1.9.0
Unidecode==0.4.18
wheel==0.24.0
